Reworked authentication token fetching to hopefully fix issues with deployed code
